requirements.txt
opencv-python
torch
transformers
Pillow
tqdm
open3d
pycolmap
matplotlib
accelerate